[macOS] Move more methods to base DisplayServer.
